+++ src/Schema/Directives/ConvertEmptyStringsToNullDirective.php
@@ -4,12 +4,10 @@ namespace Nuwave\Lighthouse\Schema\Directives;
 
 use GraphQL\Type\Definition\ScalarType;
 use Nuwave\Lighthouse\Execution\Arguments\ArgumentSet;
-use Nuwave\Lighthouse\Execution\ResolveInfo;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\ArgDirective;
 use Nuwave\Lighthouse\Support\Contracts\ArgSanitizerDirective;
 use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
-use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 use Nuwave\Lighthouse\Support\Utils;
 
 class ConvertEmptyStringsToNullDirective extends BaseDirective implements ArgSanitizerDirective, ArgDirective, FieldMiddleware
@@ -36,22 +34,9 @@ GRAPHQL;
         );
     }
 
-    public function handleField(FieldValue $fieldValue, \Closure $next): FieldValue
+    public function handleField(FieldValue $fieldValue): void
     {
-        $resolver = $fieldValue->getResolver();
-
-        $fieldValue->setResolver(function ($root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($resolver) {
-            $resolveInfo->argumentSet = $this->transformArgumentSet($resolveInfo->argumentSet);
-
-            return $resolver(
-                $root,
-                $resolveInfo->argumentSet->toArray(),
-                $context,
-                $resolveInfo
-            );
-        });
-
-        return $next($fieldValue);
+        $fieldValue->addArgumentSetTransformer(fn (ArgumentSet $argumentSet): ArgumentSet => $this->transformArgumentSet($argumentSet));
     }
 
     protected function transformArgumentSet(ArgumentSet $argumentSet): ArgumentSet