+++ src/Schema/Directives/WithRelationDirective.php
@@ -17,24 +17,17 @@ abstract class WithRelationDirective extends BaseDirective implements FieldMiddl
     use RelationDirectiveHelpers;
 
     /**
-     * @param  mixed  $parent the parent node
      * @param  array<string, mixed>  $args
      */
-    abstract protected function modelsLoader($parent, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): ModelsLoader;
+    abstract protected function modelsLoader(mixed $parent, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): ModelsLoader;
 
-    public function handleField(FieldValue $fieldValue, \Closure $next): FieldValue
+    public function handleField(FieldValue $fieldValue): void
     {
-        $previousResolver = $fieldValue->getResolver();
-
-        $fieldValue->setResolver(function (Model $parent, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($previousResolver) {
-            return $this
+        $fieldValue->wrapResolver(
+            fn (callable $resolver) => fn (Model $parent, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) => $this
                 ->loadRelation($parent, $args, $context, $resolveInfo)
-                ->then(static function () use ($previousResolver, $parent, $args, $context, $resolveInfo) {
-                    return $previousResolver($parent, $args, $context, $resolveInfo);
-                });
-        });
-
-        return $next($fieldValue);
+                ->then(static fn () => $resolver($parent, $args, $context, $resolveInfo))
+        );
     }
 
     /**