+++ src/Subscriptions/Directives/BroadcastDirective.php
@@ -14,6 +14,9 @@ class BroadcastDirective extends BaseDirective implements FieldMiddleware
         return /** @lang GraphQL */ <<<'GRAPHQL'
 """
 Broadcast the results of a mutation to subscribed clients.
+
+Ensure you place this after other field middleware directives that may transform the
+result to broadcast the final value.
 """
 directive @broadcast(
   """
@@ -30,11 +33,8 @@ directive @broadcast(
 GRAPHQL;
     }
 
-    public function handleField(FieldValue $fieldValue, \Closure $next): FieldValue
+    public function handleField(FieldValue $fieldValue): void
     {
-        // Ensure this is run after the other field middleware directives
-        $fieldValue = $next($fieldValue);
-
         $subscriptionField = $this->directiveArgValue('subscription');
         $shouldQueue = $this->directiveArgValue('shouldQueue');
 
@@ -43,7 +43,5 @@ GRAPHQL;
 
             return $root;
         });
-
-        return $fieldValue;
     }
 }