+++ src/Support/Contracts/FieldBuilderDirective.php
@@ -2,7 +2,7 @@
 
 namespace Nuwave\Lighthouse\Support\Contracts;
 
-use GraphQL\Type\Definition\ResolveInfo;
+use Nuwave\Lighthouse\Execution\ResolveInfo;
 
 interface FieldBuilderDirective extends Directive
 {