+++ src/Support/DriverManager.php
@@ -2,7 +2,6 @@
 
 namespace Nuwave\Lighthouse\Support;
 
-use Closure;
 use Illuminate\Container\Container as Application;
 use Nuwave\Lighthouse\Exceptions\InvalidDriverException;
 
@@ -91,7 +90,7 @@ abstract class DriverManager
     }
 
     /**
-     * Register a custom driver creator Closure.
+     * Register a custom driver creator callback.
      */
     public function extend(string $driver, \Closure $callback): self
     {