+++ tests/Integration/Schema/Directives/BuilderDirectiveTest.php
@@ -2,8 +2,8 @@
 
 namespace Tests\Integration\Schema\Directives;
 
-use GraphQL\Type\Definition\ResolveInfo;
 use Illuminate\Database\Eloquent\Builder;
+use Nuwave\Lighthouse\Execution\ResolveInfo;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 use Tests\DBTestCase;
 use Tests\Utils\Models\User;