+++ tests/Utils/Resolvers/Foo.php
@@ -2,8 +2,6 @@
 
 namespace Tests\Utils\Resolvers;
 
-use Illuminate\Support\Arr;
-
 final class Foo
 {
     public function __invoke(): string
@@ -14,15 +12,5 @@ final class Foo
     public function bar(): string
     {
         return 'foo.bar';
-    }
-
-    /**
-     * @param  array<string, mixed>  $args
-     *
-     * @see \Tests\Unit\Schema\Directives\FieldDirectiveTest
-     */
-    public function baz($root, array $args): string
-    {
-        return Arr::get($args, 'directive.0');
     }
 }